cavencity-ui/requirements.txt
pyserial=3.5
PyQt5==5.15.4
PyInstaller=4.3
